+++ app/client/screens/Reward.jsx
