+++ app/client/screens/StartScreen.jsx
